assignments/assign06-docker-compose/backend/requirements.txt
# Flask Backend Dependencies
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
redis==5.0.1
gunicorn==21.2.0
